pyproject.toml
[project]
name = "lab2-data-platform-foodhub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.128.8",
    "kafka-python>=2.3.0",
    "matplotlib>=3.9.4",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.2.13",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.5",
    "uvicorn>=0.39.0",
]
